// src/services/kobold/lib/shared-schemas/action.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An custom sheet action
 */
export interface Action {
	name: string;
	description: string;
	type: string;
	actionCost: string | null;
	baseLevel: number | null;
	autoHeighten: boolean;
	tags: string[];
	rolls: (
		| {
				name: string;
				type: "attack" | "skill-challenge";
				targetDC?: string | null;
				roll: string | null;
				allowRollModifiers: boolean;
				[k: string]: any;
		  }
		| {
				name: string;
				type: "damage";
				damageType?: string | null;
				healInsteadOfDamage?: boolean | null;
				roll: string | null;
				allowRollModifiers: boolean;
				[k: string]: any;
		  }
		| {
				name: string;
				type: "advanced-damage";
				damageType?: string | null;
				healInsteadOfDamage?: boolean | null;
				criticalSuccessRoll?: string | null;
				criticalFailureRoll?: string | null;
				successRoll?: string | null;
				failureRoll?: string | null;
				allowRollModifiers: boolean;
				[k: string]: any;
		  }
		| {
				name: string;
				type: "save";
				saveRollType?: string | null;
				saveTargetDC?: string | null;
				allowRollModifiers: boolean;
				[k: string]: any;
		  }
		| {
				name: string;
				type: "text";
				defaultText?: string | null;
				criticalSuccessText?: string | null;
				criticalFailureText?: string | null;
				successText?: string | null;
				failureText?: string | null;
				allowRollModifiers: boolean;
				extraTags: string[];
				[k: string]: any;
		  }
	)[];
	[k: string]: any;
}