// src/services/kobold/lib/shared-schemas/modifier.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A modifier is a bonus or penalty that can be applied to a roll or sheet property.
 */
export type Modifier = {
	name: string;
	isActive: boolean;
	description: string | null;
	type: string;
	[k: string]: any;
} & (
	| {
			modifierType: "roll";
			value: number | string;
			targetTags: string | null;
			[k: string]: any;
	  }
	| {
			modifierType: "sheet";
			sheetAdjustments: {
				property: string;
				operation: "+" | "-" | "=";
				value: string;
				[k: string]: any;
			}[];
			[k: string]: any;
	  }
);