// src/services/kobold/lib/shared-schemas/roll-macro.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface RollMacro {
	name: string;
	macro: string;
	[k: string]: any;
}